+++ pylxd/deprecated/api.py
@@ -19,7 +19,7 @@ from pylxd.deprecated import connection
 from pylxd.deprecated import container
 from pylxd.deprecated import hosts
 from pylxd import image
-from pylxd import network
+from pylxd.deprecated import network
 from pylxd import operation
 from pylxd import profiles
